# openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "77-Bit API",
    "description": "An API for interacting with the 77-Bit L2 blockchain.",
    "termsOfService": "https://77-bit.com/terms",
    "license": {
      "name": "License",
      "url": "https://77-bit.com/license"
    },
    "version": "1.0",
    "x-logo": {
      "url": "https://pbs.twimg.com/profile_images/1642220143915696128/x3yCWwbE_400x400.jpg",
      "altText": "77-Bit logo"
    }
  },
  "servers": [
    {
      "url": "https://testnet-api.77-bit.com",
      "description": "Testnet server (Goerli)."
    },
    {
      "url": "https://api.77-bit.com",
      "description": "Mainnet server (Ethereum)."
    }
  ],
  "paths": {
...
    "/api/v1/vaults/withdraw": {
      "post": {
        "tags": [
          "Withdraw"
        ],
        "summary": "Withdraw Asset",
        "description": "This endpoint allows for withdrawing assets from SevenSevenBit.",
        "operationId": "Withdraw",
        "requestBody": {
          "description": "The withdraw request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the details of the withdraw operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "The withdraw request was invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oauth2": [
              "write:vaults"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataAvailabilityModes": {
        "enum": [
          "ZkRollup",
          "Validium"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "VaultDto": {
        "type": "object",
        "properties": {
          "vaultId": {
            "title": "Vault ID",
            "type": "string",
            "description": "The ID of the vault.",
            "format": "uuid"
          },
          "vaultChainId": {
            "title": "Vault Chain ID",
            "type": "string",
            "description": "The StarkEx ID of the vault."
          },
          "assetSymbol": {
            "title": "Asset Symbol",
            "type": "string",
            "description": "The symbol of the asset associated with the vault.",
            "format": "string",
            "nullable": true
          },
          "tokenId": {
            "title": "Token ID",
            "type": "string",
            "description": "The token id of the asset associated with the vault, if the asset is an ERC721 or ERC1155.",
            "format": "hex",
            "nullable": true
          },
          "mintingBlob": {
            "title": "Minting Blob",
            "type": "string",
            "description": "The minting blob of the asset associated with the vault, if the asset is a Mintable ERC20, ERC721 or ERC1155.",
            "format": "hex",
            "nullable": true
          },
          "assetStarkExId": {
            "title": "Asset StarkEx ID",
            "type": "string",
            "description": "The StarkEx ID of the asset associated with the vault.",
            "format": "hex",
            "nullable": true
          },
          "userStarkKey": {
            "title": "STARK Key",
            "type": "string",
            "description": "The STARK key of the user associated with the vault.",
            "format": "hex",
            "nullable": true
          },
          "availableBalance": {
            "title": "Available Balance",
            "type": "string",
            "description": "The available balance of the vault."
          },
          "accountingBalance": {
            "title": "Accounting Balance",
            "type": "string",
            "description": "The accounting balance of the vault."
          },
          "dataAvailabilityMode": {
            "$ref": "#/components/schemas/DataAvailabilityModes"
          }
        },
        "additionalProperties": false
      },
      "WithdrawDetailsDto": {
        "type": "object",
        "properties": {
          "vault": {
            "$ref": "#/components/schemas/VaultDto"
          },
          "withdrawFunction": {
            "title": "Smart Contract Address",
            "type": "string",
            "description": "The withdraw function to use on-chain.",
            "format": "string",
            "nullable": true
          },
          "starkKey": {
            "title": "User's STARK Key",
            "type": "string",
            "description": "The user's public STARK key",
            "format": "hex",
            "nullable": true
          },
          "assetType": {
            "title": "Asset type",
            "type": "string",
            "description": "The asset type identifier.",
            "format": "hex",
            "nullable": true
          },
          "tokenId": {
            "title": "Token Id",
            "type": "string",
            "description": "The token Id for ERC-721 and ERC-1155 assets.",
            "format": "hex",
            "nullable": true
          },
          "mintingBlob": {
            "title": "Minting blob",
            "type": "string",
            "description": "The minting blob for Mintable ERC-20, ERC-721 and ERC-1155 assets.",
            "format": "hex",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WithdrawModel": {
        "required": [
          "amount",
          "vaultId"
        ],
        "type": "object",
        "properties": {
          "vaultId": {
            "title": "Vault ID",
            "type": "string",
            "description": "The unique identifier of the vault to withdraw from.",
            "format": "uuid"
          },
          "amount": {
            "title": "Amount",
            "type": "string",
            "description": "The amount of the asset to be withdrawn, in unquantized form."
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://sevensevenbit-dev.eu.auth0.com/oauth/token",
            "scopes": {
              "write:users": "Access user write operations.",
              "read:users": "Access user read operations.",
              "write:assets": "Access asset write operations.",
              "read:assets": "Access asset read operations.",
              "mint:assets": "Access mint operations.",
              "read:vaults": "Access vault read operations.",
              "write:transfers": "Access transfer write operations.",
              "write:settlements": "Access settlement write operations.",
              "read:transactions": "Access tarnsaction read operations."
            }
          }
        }
      }
    }
  },
...
}